requirements.txt
flask
pandas
prophet
xgboost
scikit-learn
sqlalchemy
requests
python-dotenv
